+++ game/gameanimationdialog.cpp
@@ -12,7 +12,7 @@
 
 GameAnimationDialog::GameAnimationDialog(quint32 frameTime)
     : GameAnimation (frameTime),
-      textSpeed(100 / static_cast<qint32>(Settings::getAnimationSpeed()))
+      textSpeed(100 / Settings::getAnimationSpeed())
 {
     Mainapp* pApp = Mainapp::getInstance();
     this->moveToThread(pApp->getWorkerthread());
@@ -257,7 +257,7 @@ void GameAnimationDialog::setFinishDelay(qint32 valueMs)
 
 void GameAnimationDialog::setTextSpeed(qint32 speed)
 {
-    textSpeed = speed / static_cast<qint32>(Settings::getAnimationSpeed());
+    textSpeed = speed / Settings::getAnimationSpeed();
 }
 
 void GameAnimationDialog::restart()
